fix: Support different source sets
