+++ playground/sandbox-compose/src/main/kotlin/com/sdds/playground/sandbox/core/compose/ComponentViewModel.kt
@@ -3,7 +3,7 @@ package com.sdds.playground.sandbox.core.compose
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sdds.compose.uikit.style.Style
-import com.sdds.playground.sandbox.ThemeContainer.composeTheme
+import com.sdds.playground.sandbox.composeTheme
 import com.sdds.playground.sandbox.core.ThemeManager
 import com.sdds.playground.sandbox.core.integration.ComposeStyleProvider
 import com.sdds.playground.sandbox.core.integration.component.ComponentKey
