+++ playground/sandbox-compose/src/main/kotlin/com/sdds/playground/sandbox/core/vs/ComponentFragment.kt
@@ -19,9 +19,9 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import com.sdds.playground.sandbox.R
 import com.sdds.playground.sandbox.Theme
-import com.sdds.playground.sandbox.ThemeContainer.viewTheme
 import com.sdds.playground.sandbox.core.integration.component.ComponentKey
 import com.sdds.playground.sandbox.databinding.FragmentComponentScaffoldBinding
+import com.sdds.playground.sandbox.viewTheme
 import com.sdds.testing.vs.UiState
 import com.sdds.uikit.Divider
 import com.sdds.uikit.FrameLayout
