+++ playground/sandbox-compose/src/starApp/kotlin/com/sdds/playground/sandbox/Theme.kt
@@ -1,40 +1,30 @@
 package com.sdds.playground.sandbox
 
-import androidx.annotation.StyleRes
-import androidx.compose.runtime.Composable
-import com.sdds.playground.sandbox.core.integration.component.ComponentsProviderCompose
-import com.sdds.playground.sandbox.core.integration.component.ComponentsProviderView
 import com.sdds.playground.sandbox.plasma.stards.integration.StarDsComposeComponents
 import com.sdds.playground.sandbox.plasma.stards.integration.StarDsThemeWrapper
 import com.sdds.playground.sandbox.plasma.stards.integration.StarDsViewComponents
 
-internal enum class Theme(
-    val compose: ThemeInfoCompose,
-    val view: ThemeInfoView,
-) {
-    PlasmaStarDs(
-        compose = ThemeInfoCompose(
-            components = StarDsComposeComponents,
-            themeWrapper = { StarDsThemeWrapper(it) },
-        ),
-        view = ThemeInfoView(
-            components = StarDsViewComponents,
-            themeRes = com.sdkit.star.designsystem.R.style.Sdkit_AppTheme,
-        ),
-    ),
+internal enum class Theme {
+    PlasmaStarDs,
     ;
-
-    internal class ThemeInfoCompose(
-        val components: ComponentsProviderCompose,
-        val themeWrapper: @Composable (@Composable () -> Unit) -> Unit,
-    )
-
-    internal class ThemeInfoView(
-        val components: ComponentsProviderView,
-        @StyleRes val themeRes: Int,
-    )
 
     internal companion object {
         val Default = PlasmaStarDs
     }
 }
+
+internal val composeThemes: List<ThemeInfoCompose> = listOf(
+    ThemeInfoCompose(
+        theme = Theme.PlasmaStarDs,
+        components = StarDsComposeComponents,
+        themeWrapper = { StarDsThemeWrapper(it) },
+    ),
+)
+
+internal val viewThemes: List<ThemeInfoView> = listOf(
+    ThemeInfoView(
+        theme = Theme.PlasmaStarDs,
+        components = StarDsViewComponents,
+        themeRes = com.sdkit.star.designsystem.R.style.Sdkit_AppTheme,
+    ),
+)
