+++ playground/sandbox-compose/src/starApp/kotlin/com/sdds/playground/sandbox/core/compose/ComponentScaffold.kt
@@ -28,6 +28,7 @@ import androidx.compose.ui.window.DialogWindowProvider
 import com.sdds.compose.uikit.style.Style
 import com.sdds.playground.sandbox.R
 import com.sdds.playground.sandbox.SandboxTheme
+import com.sdds.playground.sandbox.composeTheme
 import com.sdds.playground.sandbox.core.ThemeManager
 import com.sdds.playground.sandbox.core.integration.component.ComponentKey
 import com.sdkit.star.designsystem.theme.StarDsTheme
@@ -71,11 +72,12 @@ internal fun <State : UiState, S : Style> ComponentScaffold(
             contentAlignment = Alignment.Center,
         ) {
             val currentTheme by themeManager.currentTheme.collectAsState()
-            currentTheme.compose.themeWrapper {
+            val themeInfo = composeTheme(currentTheme)
+            themeInfo.themeWrapper {
                 val uiState by viewModel.uiState.collectAsState()
                 component(
                     uiState,
-                    currentTheme.compose.components.get<String, S>(key).styleProvider.style(uiState.variant),
+                    themeInfo.components.get<String, S>(key).styleProvider.style(uiState.variant),
                 )
             }
         }
